Game screen redesign
