+++ app/src/main/java/com/shegs/artreasurehunt/navigation/NestedNavItem.kt
@@ -8,5 +8,7 @@ sealed class NestedNavItem(val route: String, val label: String? = null, val ico
     object ARCameraScreen : NestedNavItem(route = "ar_screen")
     object SignUpScreen : NestedNavItem(route = "sign_up_screen")
 
+    object GameScreen : NestedNavItem(route = "game_screen")
+
     object MapScreen : NestedNavItem(route = "map_screen")
 }
